Use ID instead of UUID
"%s/uuid/id/g" & "%s/UUID/ID/g"

A recent merge (#53) to node-bleacon shortened the uuid to not include the
hex encoded major/minor numbers suffix anymore. This way, the uuid now
matches the one in the estimote apps and web interfaces now, but is not
sticker specific anymore.

The node-bleacon library checks for estimote vendor in the first place, so
we do not need to check the uuid again. The ID should be enough.